Create Restricter class
